refactor: remove unused controllers, models, and utility functions for cleaner codebase
