+++ dev-data/data/import-dev-data.js
@@ -1,7 +1,7 @@
 const fs = require('fs');
 const mongoose = require('mongoose');
 const dotenv = require('dotenv');
-const Tour = require('./../../models/tourModel');
+const Tour = require('../../src/models/tourModel');
 
 
 dotenv.config({ path: './config.env' });
