+++ src/controllers/handlerFactory.js
@@ -1,4 +1,4 @@
-const catchAsync = require('./../utils/catchAsync');
+const catchAsync = require('../utils/catchAsync');
 const AppError = require('./../utils/appError');
 const AppFeatures = require('../utils/apiFeatures');
